+++ devtools/Makefile
@@ -55,7 +55,7 @@ $(GOLANGCI_LINT): $(mkfile_dir)/install-golangci-lint.sh
 	@bash $(mkfile_dir)/install-golangci-lint.sh $(TOOLS_DESTDIR) $(GOLANGCI_LINT_HASHSUM)
 
 # Install the runsim binary with a temporary workaround of entering an outside
-# directory as the "go get" command ignores the -mod option and will polute the
+# directory as the "go get" command ignores the -mod option and will pollute the
 # go.{mod, sum} files.
 # 
 # ref: https://github.com/golang/go/issues/30515
